+++ src/form-elements/FormElementBSB.tsx
@@ -161,6 +161,7 @@ function FormElementBSB({
             isInputButton
             value={value}
             validationMessage={validationMessage}
+            lookupButtonConfig={element.lookupButton}
           />
         </div>
         {(((isDirty || displayValidationMessage) &&