+++ src/form-elements/FormElementEmail.tsx
@@ -64,6 +64,7 @@ function FormElementEmail({
             isInputButton
             value={value}
             validationMessage={validationMessage}
+            lookupButtonConfig={element.lookupButton}
           />
         </div>