+++ src/form-elements/FormElementTime.tsx
@@ -104,6 +104,7 @@ function FormElementTime({
             isInputButton
             value={value}
             validationMessage={validationMessage}
+            lookupButtonConfig={element.lookupButton}
           />
         </div>